+++ merged.txt
@@ -22,6 +22,7 @@
     "@types/react-dom": "^18.3.0",
     "bootstrap": "^5.3.3",
     "lodash": "^4.17.21",
+    "query-string": "^9.0.0",
     "react": "^18.3.1",
     "react-bootstrap": "^2.10.2",
     "react-datepicker": "^6.9.0",
@@ -251,14 +252,8 @@ import { formatTime, parseTimeString } from "./utils/timeUtils";
 import useFormState from "./hooks/useFormState";
 import useUrlParams from "./hooks/useUrlParams";
 import { getDefaultDate } from "./utils/dateUtils";
-import { parseISO, isValid } from "date-fns";
-
-const formFieldsConfig = [
-  { name: "departureTime", defaultValue: "11:00" },
-  { name: "drivingTime", defaultValue: "45" },
-  { name: "arriveEarly", defaultValue: "30" },
-  { name: "snackTime", defaultValue: "5" },
-];
+import { parseISO, isValid, isAfter } from "date-fns";
+import formFieldsConfig from "./config/formFieldsConfig";
 
 const FlightFormLogic = ({ onCalculate }) => {
   const navigate = useNavigate();
@@ -347,9 +342,16 @@ const FlightFormLogic = ({ onCalculate }) => {
   ]);
 
   const handleDateChange = (date) => {
-    setSelectedDate(date);
-    setIsInitialLoad(false);
-    console.log(`Date changed to: ${date}`);
+    const now = new Date();
+    if (isAfter(date, now)) {
+      setSelectedDate(date);
+      setIsInitialLoad(false);
+      console.log(`Future date selected: ${date}`);
+    } else {
+      setSelectedDate(date);
+      setIsInitialLoad(false);
+      console.log(`Past date or today selected: ${date}`);
+    }
   };
 
   const handleReset = () => {
@@ -553,6 +555,17 @@ const TimeInput = ({ label, value, onChange, readOnly = false }) => (
 export default TimeInput;
 
 
+// src/config/formFieldsConfig.js
+const formFieldsConfig = [
+  { name: "departureTime", defaultValue: "11:00" },
+  { name: "drivingTime", defaultValue: "45" },
+  { name: "arriveEarly", defaultValue: "30" },
+  { name: "snackTime", defaultValue: "5" },
+];
+
+export default formFieldsConfig;
+
+
 // src/hooks/useFormField.js
 import { useState, useEffect } from "react";
 import { useNavigate, useLocation } from "react-router-dom";
@@ -665,10 +678,9 @@ export default useFormState;
 
 
 // src/hooks/useUrlParams.js
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 import { useNavigate, useLocation } from "react-router-dom";
-import { parseISO, isValid } from "date-fns";
-import { debounce } from "lodash";
+import { parseISO, isValid } from "date-fns"; // Import parseISO and isValid
 
 const useUrlParams = (
   formValues,
@@ -679,13 +691,10 @@ const useUrlParams = (
 ) => {
   const navigate = useNavigate();
   const location = useLocation();
-
-  const debouncedNavigate = debounce((params) => {
-    navigate({ search: params.toString() }, { replace: true });
-  }, 300); // Adjust debounce delay as necessary
+  const [paramsUpdated, setParamsUpdated] = useState(false);
 
   useEffect(() => {
-    if (!isInitialLoad) {
+    if (!isInitialLoad && !paramsUpdated) {
       const params = new URLSearchParams(location.search);
       let shouldNavigate = false;
 
@@ -725,7 +734,8 @@ const useUrlParams = (
       }
 
       if (shouldNavigate) {
-        debouncedNavigate(params);
+        setParamsUpdated(true); // Prevent immediate re-update
+        navigate({ search: params.toString() }, { replace: true });
       }
     }
   }, [
@@ -735,15 +745,17 @@ const useUrlParams = (
     formFieldsConfig,
     defaultDate,
     location.search,
-    debouncedNavigate,
+    navigate,
+    paramsUpdated,
   ]);
 
-  // Cleanup debounce on unmount
+  // Reset paramsUpdated after some time to allow further updates
   useEffect(() => {
-    return () => {
-      debouncedNavigate.cancel();
-    };
-  }, [debouncedNavigate]);
+    if (paramsUpdated) {
+      const timer = setTimeout(() => setParamsUpdated(false), 300); // Adjust the delay as necessary
+      return () => clearTimeout(timer);
+    }
+  }, [paramsUpdated]);
 };
 
 export default useUrlParams;
